# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Web API Express",
        "description": "My WebApi using Node",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/deleteUser": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Elimina un usuario",
                "description": "Elimina un usuario con el uso del correo como identificador",
                "operationId": "deleteUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nuevo jugador",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DeleteUser"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        }
    },
    "definitions":{
...
        "DeleteUser":{
            "type": "object",
            "properties":{
                "email":{
                    "type": "string"
                }
            }
        }
    }
}